LayoutTests/fast/text/FontFaceSet-status-after-style-update.html

<!DOCTYPE html>
<html>
<head id="head">
<script>
if (window.internals) {
    internals.invalidateFontCache();
    internals.clearMemoryCache();
}
</script>
<style id="style">
@font-face {
    font-family: "WebFont";
    src: url("../../resources/Ahem.ttf") format("truetype");
}
</style>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/ui-helper.js"></script>
</head>
<body>
<script>    

jsTestIsAsync = true;
let fonts = document.fonts;
(async () => {
    description("This function makes sure document.fonts.status gets updated after style changes.");
    shouldBeEqualToString("fonts.status", "loaded");
    debug("Loading font");
    fonts.values().next().value.load();
    shouldBeEqualToString("fonts.status", "loading");
    debug("Deleting style");
    document.getElementById("head").removeChild(document.getElementById("style"));

    await UIHelper.renderingUpdate();

    shouldBeEqualToString("fonts.status", "loaded");
    finishJSTest();
})();

</script>
</body>
</html>
